requirements.txt
pymongo==3.6.1
pyyaml==3.12
uwsgi==2.0.17
